refactor: move batching managers to a new namespace
