+++ Testing/VelaptorTests/Batching/FontGlyphBatchingManagerTests.cs
@@ -1,8 +1,8 @@
-﻿// <copyright file="TextureBatchingManagerTests.cs" company="KinsonDigital">
+﻿// <copyright file="FontGlyphBatchingManagerTests.cs" company="KinsonDigital">
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace VelaptorTests.Services;
+namespace VelaptorTests.Batching;
 
 using System;
 using System.Collections.Generic;
@@ -14,6 +14,7 @@ using Carbonate.Core.UniDirectional;
 using Carbonate.NonDirectional;
 using Carbonate.UniDirectional;
 using FluentAssertions;
+using Helpers;
 using Moq;
 using Velaptor;
 using Velaptor.Exceptions;
@@ -21,24 +22,24 @@ using Velaptor.Factories;
 using Velaptor.Graphics;
 using Velaptor.OpenGL.Batching;
 using Velaptor.ReactableData;
-using Velaptor.Services;
+using Velaptor.Batching;
 using Xunit;
 
 /// <summary>
-/// Tests the <see cref="TextureBatchingManager"/> class.
+/// Tests the <see cref="FontGlyphBatchingManager"/> class.
 /// </summary>
-public class TextureBatchingManagerTests
+public class FontGlyphBatchingManagerTests
 {
-    private readonly Mock<IDisposable> mockUnsubscriber;
-    private readonly Mock<IPushReactable> mockPushReactable;
-    private readonly Mock<IPushReactable<BatchSizeData>> mockBatchSizeReactable;
     private readonly Mock<IReactableFactory> mockReactableFactory;
-    private IReceiveReactor<BatchSizeData>? reactor;
+    private readonly Mock<IPushReactable<BatchSizeData>> mockBatchSizeReactable;
+    private readonly Mock<IPushReactable> mockPushReactable;
+    private readonly Mock<IDisposable> mockUnsubscriber;
+    private IReceiveReactor<BatchSizeData>? batchSizeReactor;
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="TextureBatchingManagerTests"/> class.
+    /// Initializes a new instance of the <see cref="FontGlyphBatchingManagerTests"/> class.
     /// </summary>
-    public TextureBatchingManagerTests()
+    public FontGlyphBatchingManagerTests()
     {
         this.mockUnsubscriber = new Mock<IDisposable>();
 
@@ -46,7 +47,11 @@ public class TextureBatchingManagerTests
 
         this.mockBatchSizeReactable = new Mock<IPushReactable<BatchSizeData>>();
         this.mockBatchSizeReactable.Setup(m => m.Subscribe(It.IsAny<IReceiveReactor<BatchSizeData>>()))
-            .Callback<IReceiveReactor<BatchSizeData>>(reactorObj => this.reactor = reactorObj)
+            .Callback<IReceiveReactor<BatchSizeData>>(reactorObj =>
+            {
+                reactorObj.Should().NotBeNull();
+                this.batchSizeReactor = reactorObj;
+            })
             .Returns(this.mockUnsubscriber.Object);
 
         this.mockReactableFactory = new Mock<IReactableFactory>();
@@ -61,7 +66,7 @@ public class TextureBatchingManagerTests
         // Arrange & Act
         var act = () =>
         {
-            _ = new TextureBatchingManager(null);
+            _ = new FontGlyphBatchingManager(null);
         };
 
         // Assert
@@ -73,15 +78,14 @@ public class TextureBatchingManagerTests
     [Fact]
     public void Ctor_WhenReceivingBatchSizePushNotification_CreatesBatchItemList()
     {
-        // Arrange
+        // Arrange & Act
         var mockMessage = new Mock<IMessage<BatchSizeData>>();
         mockMessage.Setup(m => m.GetData(It.IsAny<Action<Exception>?>()))
             .Returns(new BatchSizeData { BatchSize = 4u });
 
-        var sut = CreateSystemUnderTest();
+        var sut = CreateService();
 
-        // Act
-        this.reactor.OnReceive(mockMessage.Object);
+        this.batchSizeReactor.OnReceive(mockMessage.Object);
 
         // Assert
         sut.BatchItems.Should().HaveCount(4);
@@ -91,12 +95,12 @@ public class TextureBatchingManagerTests
     public void Ctor_WhenReactableUnsubscribes_InvokesUnsubscriber()
     {
         // Arrange
-        _ = CreateSystemUnderTest();
+        _ = CreateService();
 
         // Act
-        this.reactor.OnUnsubscribe();
+        this.batchSizeReactor.OnUnsubscribe();
+        this.batchSizeReactor.OnUnsubscribe();
 
-        this.reactor.OnUnsubscribe();
         // Assert
         this.mockUnsubscriber.Verify(m => m.Dispose());
     }
@@ -105,7 +109,7 @@ public class TextureBatchingManagerTests
     public void Ctor_WhenBatchSizeNotificationHasAnIssue_ThrowsException()
     {
         // Arrange
-        var expectedMsg = $"There was an issue with the '{nameof(TextureBatchingManager)}.Constructor()' subscription source";
+        var expectedMsg = $"There was an issue with the '{nameof(FontGlyphBatchingManager)}.Constructor()' subscription source";
         expectedMsg += $" for subscription ID '{NotificationIds.BatchSizeSetId}'.";
 
         this.mockBatchSizeReactable.Setup(m => m.Subscribe(It.IsAny<IReceiveReactor<BatchSizeData>>()))
@@ -113,17 +117,17 @@ public class TextureBatchingManagerTests
             {
                 reactorObj.Should().NotBeNull("it is required for unit testing.");
 
-                this.reactor = reactorObj;
+                this.batchSizeReactor = reactorObj;
             });
 
         var mockMessage = new Mock<IMessage<BatchSizeData>>();
         mockMessage.Setup(m => m.GetData(null))
             .Returns<Action<Exception>?>(_ => null);
 
-        _ = CreateSystemUnderTest();
+        _ = CreateService();
 
         // Act
-        var act = () => this.reactor.OnReceive(mockMessage.Object);
+        var act = () => this.batchSizeReactor.OnReceive(mockMessage.Object);
 
         // Assert
         act.Should().Throw<PushNotificationException>()
@@ -136,18 +140,20 @@ public class TextureBatchingManagerTests
     public void BatchItems_WhenSettingValue_ReturnsCorrectResult()
     {
         // Arrange
-        var batchItem1 = new TextureBatchItem(
+        var batchItem1 = new FontGlyphBatchItem(
             new RectangleF(7, 8, 9, 10),
             new RectangleF(3, 4, 5, 6),
+            'g',
             2,
             1,
             Color.FromArgb(12, 13, 14, 15),
             RenderEffects.None,
             11,
-            18);
-        var batchItem2 = new TextureBatchItem(
-            new RectangleF(24, 25, 26, 27),
+            12);
+        var batchItem2 = new FontGlyphBatchItem(
             new RectangleF(20, 21, 22, 23),
+            new RectangleF(24, 25, 26, 27),
+            'g',
             19,
             18,
             Color.FromArgb(29, 30, 31, 32),
@@ -155,131 +161,129 @@ public class TextureBatchingManagerTests
             28,
             35);
 
-        var batchItems = new List<TextureBatchItem> { batchItem1, batchItem2 };
-        var expected = new ReadOnlyCollection<TextureBatchItem>(batchItems.ToReadOnlyCollection());
-        var sut = CreateSystemUnderTest();
+        var batchItems = new List<FontGlyphBatchItem> { batchItem1, batchItem2 };
+        var expected = new ReadOnlyCollection<FontGlyphBatchItem>(batchItems.ToReadOnlyCollection());
+        var service = CreateService();
 
         // Act
-        sut.BatchItems = batchItems.ToReadOnlyCollection();
-        var actual = sut.BatchItems;
+        service.BatchItems = batchItems.ToReadOnlyCollection();
+        var actual = service.BatchItems;
 
         // Assert
-        actual.Should().BeEquivalentTo(expected.ToArray());
+        AssertExtensions.ItemsEqual(expected.ToArray(), actual.ToArray());
+        AssertExtensions.ItemsEqual(expected.ToArray(), actual.ToArray());
     }
     #endregion
 
     #region Method Tests
     [Fact]
-    public void Add_WhenBatchIsFull_RaisesBatchFilledEvent()
+    public void Add_WhenBatchIsFull_SendsRenderPushNotification()
     {
         // Arrange
-        var batchItem1 = new TextureBatchItem(
-            RectangleF.Empty,
-            RectangleF.Empty,
-            1,
+        var batchItem1 = new FontGlyphBatchItem(
+            default,
+            default,
+            'g',
+            0,
             0,
             Color.Empty,
             RenderEffects.None,
-            1,
+            0,
             0);
-
-        var batchItem2 = new TextureBatchItem(
-            RectangleF.Empty,
-            RectangleF.Empty,
-            2,
+        var batchItem2 = new FontGlyphBatchItem(
+            default,
+            default,
+            'g',
+            0,
             0,
             Color.Empty,
             RenderEffects.None,
-            2,
+            0,
             0);
 
         var mockMessage = new Mock<IMessage<BatchSizeData>>();
         mockMessage.Setup(m => m.GetData(It.IsAny<Action<Exception>?>()))
             .Returns(new BatchSizeData { BatchSize = 1u });
 
-        var sut = CreateSystemUnderTest();
-
-        this.reactor.OnReceive(mockMessage.Object);
-        sut.Add(batchItem1);
+        var service = CreateService();
+        this.batchSizeReactor.OnReceive(mockMessage.Object);
+        service.Add(batchItem1);
 
         // Act
-        sut.Add(batchItem2);
+        service.Add(batchItem2);
 
         // Assert
-        this.mockPushReactable.Verify(m => m.Push(NotificationIds.RenderTexturesId));
+        this.mockPushReactable.Verify(m => m.Push(NotificationIds.RenderFontsId));
     }
 
     [Fact]
     public void EmptyBatch_WhenInvoked_EmptiesAllItemsReadyToRender()
     {
         // Arrange
-        var expected = new[] { default(TextureBatchItem), default(TextureBatchItem) };
-        var batchItem1 = new TextureBatchItem(
-            RectangleF.Empty,
-            RectangleF.Empty,
-            1,
+        var batchItem1 = new FontGlyphBatchItem(
+            default,
+            default,
+            'g',
+            0,
             0,
             Color.Empty,
             RenderEffects.None,
-            1,
+            0,
             0);
-        var batchItem2 = new TextureBatchItem(
-            RectangleF.Empty,
-            RectangleF.Empty,
-            2,
+        var batchItem2 = new FontGlyphBatchItem(
+            default,
+            default,
+            'g',
+            0,
             0,
             Color.Empty,
             RenderEffects.None,
-            2,
+            0,
             0);
 
         var mockMessage = new Mock<IMessage<BatchSizeData>>();
         mockMessage.Setup(m => m.GetData(It.IsAny<Action<Exception>?>()))
             .Returns(new BatchSizeData { BatchSize = 2u });
 
-        var sut = CreateSystemUnderTest();
-
-        this.reactor.OnReceive(mockMessage.Object);
-
-        sut.Add(batchItem1);
-        sut.Add(batchItem2);
+        var service = CreateService();
+        this.batchSizeReactor.OnReceive(mockMessage.Object);
+        service.Add(batchItem1);
+        service.Add(batchItem2);
 
         // Act
-        sut.EmptyBatch();
+        service.EmptyBatch();
 
         // Assert
-        sut.BatchItems.Should().BeEquivalentTo(expected);
+        Assert.NotEqual(batchItem1, service.BatchItems[0]);
     }
 
     [Fact]
     public void EmptyBatch_WithNoItemsReadyToRender_DoesNotEmptyItems()
     {
         // Arrange
-        var batchItem1 = default(TextureBatchItem);
-        var batchItem2 = default(TextureBatchItem);
+        var batchItem1 = default(FontGlyphBatchItem);
+        var batchItem2 = default(FontGlyphBatchItem);
 
         var mockMessage = new Mock<IMessage<BatchSizeData>>();
         mockMessage.Setup(m => m.GetData(It.IsAny<Action<Exception>?>()))
             .Returns(new BatchSizeData { BatchSize = 2u });
 
-        var sut = CreateSystemUnderTest();
-
-        this.reactor.OnReceive(mockMessage.Object);
-
-        sut.BatchItems = new List<TextureBatchItem> { batchItem1, batchItem2 }.ToReadOnlyCollection();
+        var service = CreateService();
+        this.batchSizeReactor.OnReceive(mockMessage.Object);
+        service.BatchItems = new List<FontGlyphBatchItem> { batchItem1, batchItem2 }.ToReadOnlyCollection();
 
         // Act
-        sut.EmptyBatch();
+        service.EmptyBatch();
 
         // Assert
-        sut.BatchItems[0].Should().BeEquivalentTo(batchItem1);
-        sut.BatchItems[1].Should().BeEquivalentTo(batchItem2);
+        Assert.Equal(batchItem1, service.BatchItems[0]);
+        Assert.Equal(batchItem2, service.BatchItems[1]);
     }
     #endregion
 
     /// <summary>
-    /// Creates a new instance of <see cref="TextureBatchingManager"/> for the purpose of testing.
+    /// Creates a new instance of <see cref="FontGlyphBatchingManager"/> for the purpose of testing.
     /// </summary>
     /// <returns>The instance to test.</returns>
-    private TextureBatchingManager CreateSystemUnderTest() => new (this.mockReactableFactory.Object);
+    private FontGlyphBatchingManager CreateService() => new (this.mockReactableFactory.Object);
 }
