+++ Testing/VelaptorTests/Graphics/Renderers/LineRendererTests.cs
@@ -13,6 +13,7 @@ using Carbonate.NonDirectional;
 using FluentAssertions;
 using Moq;
 using Velaptor;
+using Velaptor.Batching;
 using Velaptor.Content;
 using Velaptor.Factories;
 using Velaptor.Graphics;
@@ -22,7 +23,6 @@ using Velaptor.OpenGL;
 using Velaptor.OpenGL.Batching;
 using Velaptor.OpenGL.Buffers;
 using Velaptor.OpenGL.Shaders;
-using Velaptor.Services;
 using Xunit;
 
 /// <summary>
