+++ Velaptor/Batching/TextureBatchingManager.cs
@@ -1,8 +1,8 @@
-﻿// <copyright file="LineBatchingManager.cs" company="KinsonDigital">
+// <copyright file="TextureBatchingManager.cs" company="KinsonDigital">
 // Copyright (c) KinsonDigital. All rights reserved.
 // </copyright>
 
-namespace Velaptor.Services;
+namespace Velaptor.Batching;
 
 using System;
 using System.Collections.Generic;
@@ -10,6 +10,7 @@ using System.Collections.ObjectModel;
 using System.Linq;
 using Carbonate.NonDirectional;
 using Carbonate.UniDirectional;
+using Content;
 using Exceptions;
 using Factories;
 using Guards;
@@ -17,19 +18,19 @@ using OpenGL.Batching;
 using ReactableData;
 
 /// <summary>
-/// Manages the process of batching the rendering of lines.
+/// Manages the process of batching up the rendering of <see cref="ITexture"/>s.
 /// </summary>
-internal sealed class LineBatchingManager : IBatchingManager<LineBatchItem>
+internal sealed class TextureBatchingManager : IBatchingManager<TextureBatchItem>
 {
     private readonly IDisposable unsubscriber;
     private readonly IPushReactable pushReactable;
-    private LineBatchItem[] batchItems = null!;
+    private TextureBatchItem[] batchItems = null!;
 
     /// <summary>
-    /// Initializes a new instance of the <see cref="LineBatchingManager"/> class.
+    /// Initializes a new instance of the <see cref="TextureBatchingManager"/> class.
     /// </summary>
     /// <param name="reactableFactory">Creates reactables for sending and receiving notifications with or without data.</param>
-    public LineBatchingManager(IReactableFactory reactableFactory)
+    public TextureBatchingManager(IReactableFactory reactableFactory)
     {
         EnsureThat.ParamIsNotNull(reactableFactory);
 
@@ -48,11 +49,11 @@ internal sealed class LineBatchingManager : IBatchingManager<LineBatchItem>
                 if (batchSize is null)
                 {
                     throw new PushNotificationException(
-                        $"{nameof(LineBatchingManager)}.Constructor()",
+                        $"{nameof(TextureBatchingManager)}.Constructor()",
                         NotificationIds.BatchSizeSetId);
                 }
 
-                var items = new List<LineBatchItem>();
+                var items = new List<TextureBatchItem>();
 
                 for (var i = 0u; i < batchSize; i++)
                 {
@@ -65,20 +66,23 @@ internal sealed class LineBatchingManager : IBatchingManager<LineBatchItem>
     }
 
     /// <inheritdoc/>
-    public ReadOnlyCollection<LineBatchItem> BatchItems
+    public ReadOnlyCollection<TextureBatchItem> BatchItems
     {
         get => new (this.batchItems);
         set => this.batchItems = value.ToArray();
     }
 
-    /// <inheritdoc/>
-    public void Add(in LineBatchItem line)
+    /// <summary>
+    /// Adds the given <paramref name="item"/> to the batch.
+    /// </summary>
+    /// <param name="item">The item to be added.</param>
+    public void Add(in TextureBatchItem item)
     {
         var batchIsFull = this.batchItems.All(i => i.IsEmpty() is false);
 
         if (batchIsFull)
         {
-            this.pushReactable.Push(NotificationIds.RenderLinesId);
+            this.pushReactable.Push(NotificationIds.RenderTexturesId);
         }
 
         var emptyItemIndex = this.batchItems.IndexOf(i => i.IsEmpty());
@@ -88,10 +92,15 @@ internal sealed class LineBatchingManager : IBatchingManager<LineBatchItem>
             return;
         }
 
-        this.batchItems[emptyItemIndex] = line;
+        this.batchItems[emptyItemIndex] = item;
     }
 
-    /// <inheritdoc/>
+    /// <summary>
+    /// Empties the entire batch.
+    /// </summary>
+    /// <remarks>
+    /// An empty batch are items that are set NOT to render and have all values set to default.
+    /// </remarks>
     public void EmptyBatch()
     {
         for (var i = 0u; i < this.batchItems.Length; i++)
